# storage/api-docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Garage - Authentification JWT",
    "description": "API d'authentification pour les professionnels (JWT)",
    "version": "1.0.0",
    "contact": {
      "name": "TOOAUTO Garage"
    }
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Token JWT (format: Bearer {token})"
    }
  },
  "security": [],
...
  "paths": {
...
    "/auth/password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Mettre à jour le mot de passe",
        "description": "Changer le mot de passe (requiert le mot de passe actuel)",
        "operationId": "updatePassword",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["current_password", "password", "password_confirmation"],
              "properties": {
                "current_password": { "type": "string", "format": "password", "description": "Mot de passe actuel" },
                "password": { "type": "string", "format": "password", "example": "nouveau123" },
                "password_confirmation": { "type": "string", "format": "password", "example": "nouveau123" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Mot de passe mis à jour" },
          "401": { "description": "Non authentifié" },
          "422": { "description": "Mot de passe actuel incorrect ou erreur de validation" }
        }
      },
      "patch": {
        "tags": ["Auth"],
        "summary": "Mettre à jour le mot de passe",
        "operationId": "updatePasswordPatch",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["current_password", "password", "password_confirmation"],
              "properties": {
                "current_password": { "type": "string", "format": "password" },
                "password": { "type": "string", "format": "password" },
                "password_confirmation": { "type": "string", "format": "password" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Mot de passe mis à jour" },
          "401": { "description": "Non authentifié" },
          "422": { "description": "Erreur de validation" }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rafraîchir le token",
        "description": "Obtenir un nouveau token JWT. Envoyer le token actuel dans le header Authorization: Bearer {token}. Le nouveau token a la même durée de validité (ex. 24h).",
        "operationId": "refresh",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Nouveau token émis",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Token rafraîchi avec succès" },
                "token": { "type": "string" },
                "token_type": { "type": "string", "example": "bearer" },
                "expires_in": { "type": "integer", "description": "Durée de validité en secondes" },
                "expires_at": { "type": "integer", "description": "Date d'expiration en millisecondes (timestamp Unix)" }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré"
          }
        }
      }
    },
...
  },
...
}